+++ hadoop-hdds/framework/src/test/java/org/apache/hadoop/hdds/security/ssl/TestReloadingX509KeyManager.java
@@ -19,7 +19,7 @@ package org.apache.hadoop.hdds.security.ssl;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
-import org.apache.hadoop.hdds.security.x509.CertificateClientTest;
+import org.apache.hadoop.hdds.security.x509.certificate.client.CertificateClientTestImpl;
 import org.apache.ozone.test.GenericTestUtils;
 import org.apache.ozone.test.GenericTestUtils.LogCapturer;
 import org.junit.BeforeClass;
@@ -39,12 +39,12 @@ public class TestReloadingX509KeyManager {
   private final LogCapturer reloaderLog =
       LogCapturer.captureLogs(ReloadingX509KeyManager.LOG);
   private static OzoneConfiguration conf;
-  private static CertificateClientTest caClient;
+  private static CertificateClientTestImpl caClient;
 
   @BeforeClass
   public static void setUp() throws Exception {
     conf = new OzoneConfiguration();
-    caClient = new CertificateClientTest(conf);
+    caClient = new CertificateClientTestImpl(conf);
   }
 
   @Test
